Habe mir die Frechheit erlaubt, an einigen Stellen ein bisschen zu optimieren...
- Common.h für allgemeine Makros etc. hinzugefügt
- Makro SAFE_DELETE(_x_) hinzugefügt
- diverse Optimierungen
  - SAFE_DELETE anstelle von delete
  - Objekte werden möglichst nicht mit new angelegt
  - Übergabe von Daten als Parameter, nicht als Rückgabewert (einige mit new angelegte Objekte wurden in anderen Funktionen gelöscht)

+++ src/RegEx.h
@@ -111,7 +111,7 @@ class RegEx
       bool Search(const char * subject, int len = -1, int options = 0)
       {
          ClearMatchList();
-         return pcre_exec(re, pe, lastsubject = subject, slen = (len >= 0) ? len : strlen(subject), 0, options, ovector, 3*substrcount) > 0;
+         return pcre_exec(re, pe, lastsubject = subject, slen = (len >= 0) ? len : (int)strlen(subject), 0, options, ovector, 3*substrcount) > 0;
       }
 
       /////////////////////////////////
